Separate source file parsing & processing
