+++ src/test/code-sample.test.ts
@@ -1,6 +1,5 @@
-import {stripSource, applyPrefixes} from '../code-sample';
+import {stripSource, applyPrefixes, extractSamples} from '../code-sample';
 import {dedent} from '../utils';
-import {extractAsciidocSamples} from '../asciidoc';
 import {baseSample} from './common';
 
 const ASCIIDOC_PREPEND = `
@@ -48,7 +47,7 @@ const c: ABC = 'c';
 
 describe('code-sample', () => {
   test('prepend directive', () => {
-    expect(applyPrefixes(extractAsciidocSamples(ASCIIDOC_PREPEND, 'prepend', 'source.asciidoc'))).toEqual([
+    expect(applyPrefixes(extractSamples(ASCIIDOC_PREPEND, 'prepend', 'source.asciidoc'))).toEqual([
       {
         ...baseSample,
         language: 'ts',
@@ -74,7 +73,7 @@ describe('code-sample', () => {
 
   test('multiple prepend directives', () => {
     expect(
-      applyPrefixes(extractAsciidocSamples(ASCIIDOC_PREPEND_MULTIPLE, 'mpd', 'source.asciidoc')),
+      applyPrefixes(extractSamples(ASCIIDOC_PREPEND_MULTIPLE, 'mpd', 'source.asciidoc')),
     ).toEqual([
       {
         ...baseSample,
@@ -105,7 +104,7 @@ describe('code-sample', () => {
   test('prepend-subset directive', () => {
     expect(
       applyPrefixes(
-        extractAsciidocSamples(
+        extractSamples(
           dedent`
     // verifier:prepend-subset-to-following:1-3
     [[a]]
@@ -154,7 +153,7 @@ describe('code-sample', () => {
   test('prepend-subset-of-js', () => {
     expect(
       applyPrefixes(
-        extractAsciidocSamples(
+        extractSamples(
           dedent`
     // verifier:prepend-subset-to-following:1-2
     [[a]]
@@ -200,7 +199,7 @@ describe('code-sample', () => {
   test('does not prepend to -output', () => {
     expect(
       applyPrefixes(
-        extractAsciidocSamples(
+        extractSamples(
           dedent`
       // verifier:prepend-to-following
       [[hello]]
